Auto expand tree based on active page
